keyez/requirements-cpu.txt
# Core Django
Django==5.2.7
gunicorn==23.0.0
whitenoise==6.11.0

# Numerical Computing (NumPy only for Seq2Seq)
# This replaces the huge Torch library
numpy==2.2.6

# File Processing
python-docx==1.1.2
pdf2image==1.17.0
google-cloud-vision==3.9.0

# Image handling (for pdf2image)
Pillow==11.0.0

# Template Engine
Jinja2==3.1.6
MarkupSafe==3.0.3

# Utilities
asgiref==3.10.0
sqlparse==0.5.3
filelock==3.20.0
fsspec==2025.10.0
packaging==25.0
typing-extensions==4.15.0
